% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{figsave}
\alias{figsave}
\title{Save a ggplot object
This function saves a ggplot object}
\usage{
figsave(
  plot,
  filename,
  format = c("pdf", "emf", "png", "jpeg", "tiff", "svg")[1],
  width = 4,
  height = 3,
  directory = NULL,
  dpi = 300
)
}
\arguments{
\item{plot}{The plot to save.}

\item{filename}{The name of the file to save the plot to. If not specified, the plot name is used.}

\item{format}{The format to save the plot in (default: "pdf"). Options are "pdf", "emf", "png", "jpeg", "tiff", and "svg".}

\item{width}{The width of the plot in inches (default: 8).}

\item{height}{The height of the plot in inches (default: 6).}

\item{directory}{The directory to save the plot in. If not specified, the current working directory is used.}

\item{dpi}{The resolution in dots per inch (default: 300).}
}
\value{
A message indicating that the plot has been saved.
}
\description{
Saves a ggplot object
}
