% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{get_args}
\alias{get_args}
\title{Retrieve Arguments from Function Call}
\usage{
get_args()
}
\value{
A list with the arguments of the function call.
}
\description{
Retrieve arguments from a function call to the current function.
}
\details{
This function retrieves the arguments from a function call to the current function.
}
