% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{get_elements}
\alias{get_elements}
\title{Get Elements from a Nested List
This function retrieves all elements with a specific name from a nested list.}
\usage{
get_elements(x, element)
}
\arguments{
\item{x}{A nested list.}

\item{element}{A character string representing the name of the element to retrieve.}
}
\value{
A list containing all elements with the specified name.
}
\description{
Retrieves elements from a nested list.
note that this code is retrieved from https://stackoverflow.com/questions/64578972/pull-all-elements-with-specific-name-from-a-nested-list
}
