% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{ifelse.unit}
\alias{ifelse.unit}
\title{If-else for Data Frames with Units}
\usage{
ifelse.unit(cond, yes, no)
}
\arguments{
\item{cond}{A logical vector or a condition.}

\item{yes}{A value to return if the condition is TRUE.}

\item{no}{A value to return if the condition is FALSE.}
}
\value{
A vector with the same class as yes, with units if yes has units.
}
\description{
If-else for data frames with units.
}
\details{
This function is a modified version of ifelse for data frames with units. __DEPRECATED
}
