% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{make_gas_label}
\alias{make_gas_label}
\title{Create a Label for Gases with the Right Formatting
This function creates a label for gases with the right formatting, using uppercase letters for the gas name and subscripts for the numbers.}
\usage{
make_gas_label(gas)
}
\arguments{
\item{gas}{A character string representing the gas name.}
}
\value{
A character string representing the formatted gas label.
}
\description{
Creates a label for gases with the right formatting.
}
