% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{paste_}
\alias{paste_}
\title{Concatenate Strings with an Underscore}
\usage{
paste_(...)
}
\arguments{
\item{...}{Character strings to be concatenated.}
}
\value{
A character string with the concatenated strings separated by an underscore.
}
\description{
Concatenates strings with an underscore.
}
\details{
This function concatenates character strings with an underscore.
}
