% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{rollmean}
\alias{rollmean}
\title{Calculate a Rolling Mean with a Minimum Number of Observations
This function calculates a rolling mean for a numeric vector with a specified width and a minimum number of observations required to compute the mean.}
\usage{
rollmean(x, minobs = width/2, width, align = c("left", "right")[1])
}
\arguments{
\item{x}{A numeric vector.}

\item{minobs}{An integer specifying the minimum number of observations required to compute the mean (default: width/2).}

\item{width}{An integer specifying the width of the rolling mean window.}

\item{align}{A character string specifying the alignment of the rolling mean ("left" or "right", default: "left").}
}
\value{
A numeric vector representing the rolling mean with the same length as the input vector, with NA values where the minimum number of observations is not met.
}
\description{
Calculates a rolling mean with a minimum number of observations.
}
