% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{rowMeans.unit}
\alias{rowMeans.unit}
\title{Row Means for Data Frames with Units}
\usage{
rowMeans.unit(data, ...)
}
\arguments{
\item{data}{A data frame with units.}

\item{...}{Additional arguments passed to the `rowMeans` function.}
}
\value{
A numeric vector with the row means of the data frame.
}
\description{
Row means for data frames with units.
}
\details{
This function calculates row means for data frames with units.
}
