% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{rowMins.unit}
\alias{rowMins.unit}
\title{Row Mins for Data Frames with Units}
\usage{
rowMins.unit(data, ...)
}
\arguments{
\item{data}{A data frame with units.}

\item{...}{Additional arguments passed to the `pmin` function.}
}
\value{
A numeric vector with the row minimums of the data frame.
}
\description{
Row minimums for data frames with units.
}
\details{
This function calculates row minimums for data frames with units.
}
