% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{rowSums.unit}
\alias{rowSums.unit}
\title{Row Sums for Data Frames with Units}
\usage{
rowSums.unit(data, ...)
}
\arguments{
\item{data}{A data frame with units.}

\item{...}{Additional arguments passed to the `rowSums` function.}
}
\value{
A numeric vector with the row sums of the data frame.
}
\description{
Row sums for data frames with units.
}
\details{
This function calculates row sums for data frames with units.
}
