% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{RSfigsave}
\alias{RSfigsave}
\title{Save the figure currently displayed in RStudio
This function saves the figure displayed in RStudio in the specified format.}
\usage{
RSfigsave(
  filename = NULL,
  format = c("png", "jpeg", "tiff", "svg")[1],
  width = 4,
  height = 3,
  directory = NULL,
  dpi = 300
)
}
\arguments{
\item{filename}{The name of the file to save the figure to. If not specified, the figure name is used. do not use a file extension, it will be added automatically.}

\item{format}{The format to save the figure in (default: "pdf"). Options are "pdf", "png", "jpeg", "tiff", and "svg".}

\item{width}{The width of the figure in inches (default: 8).}

\item{height}{The height of the figure in inches (default: 6).}

\item{directory}{The directory to save the figure in. If not specified, the current working directory is used.}

\item{dpi}{The resolution in dots per inch (default: 300).}
}
\value{
A message indicating that the figure has been saved.
}
\description{
Saves a figure in RStudio.
}
