% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{safe_drop_units}
\alias{safe_drop_units}
\title{Drop Units from a Data Frame or a Vector}
\usage{
safe_drop_units(x)
}
\arguments{
\item{x}{A data frame or a vector with units.}
}
\value{
A data frame or a vector without units.
}
\description{
Drops units from a data frame or a vector.
}
\details{
This function drops units from a data frame or a vector.

If x is a data frame, all columns with units are dropped.
}
