% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{safe_unit_install}
\alias{safe_unit_install}
\title{Install a Unit if Not Already Installed}
\usage{
safe_unit_install(
  symbol = character(0),
  def = character(0),
  name = character(0)
)
}
\arguments{
\item{symbol}{A character string representing the unit symbol.}

\item{def}{A character string representing the unit definition.}

\item{name}{A character string representing the unit name.}
}
\description{
Installs a unit if it is not already installed.
}
\details{
This function installs a unit if it is not already installed.

If the unit is already installed, it is removed first.
}
