% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{se}
\alias{se}
\title{Calculate the Standard Error of the Mean}
\usage{
se(x, na.rm = TRUE, round = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical, if TRUE, NA values are removed before calculation.}

\item{round}{Logical, if TRUE, the result is rounded to 3 significant digits.}
}
\value{
A numeric value representing the standard error of the mean.
}
\description{
Calculates the standard error of the mean.
}
\details{
This function calculates the standard error of the mean for a numeric vector.
}
