% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{se_median}
\alias{se_median}
\title{Calculate the Standard Error of the Median}
\usage{
se_median(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the standard error of the median.
}
\description{
Calculates the standard error of the median.
}
\details{
This function calculates the standard error of the median for a numeric vector.
}
