% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{set_names_and_units}
\alias{set_names_and_units}
\title{Set Names and Units for a Data Frame
This function sets names and units for a data frame based on a specified format.}
\usage{
set_names_and_units(
  df,
  split = "[\\\\[\\\\]]",
  comment = "\\\\s*\\\\([^\\\\)]+\\\\)",
  custom_units = NULL
)
}
\arguments{
\item{df}{A data frame with column names containing labels and units.}

\item{split}{A regular expression to split the column names into labels and units (default: "[\\[\\]]").}

\item{comment}{A regular expression to remove comments from the column names (default: "\\s*\\([^\\)]+\\)").}

\item{custom_units}{A character vector of custom units to install (default: NULL).}
}
\value{
A data frame with updated column names and units.
}
\description{
Sets names and units for a data frame.
}
\details{
The function removes comments from the column names, splits the names into labels and units, and sets the units for numeric columns. If a unit is not specified, it defaults to "1". It also installs any custom units specified in the `custom_units` argument.
}
