% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{transfer_units}
\alias{transfer_units}
\title{Transfer Units from One Object to Another}
\usage{
transfer_units(recipient, donor, silent = TRUE)
}
\arguments{
\item{recipient}{An object to which the units will be transferred.}

\item{donor}{An object from which the units will be taken.}

\item{silent}{Logical, if TRUE, no error is thrown if the recipient already has units.}
}
\value{
An object with the units of the donor.
}
\description{
Transfers units from one object to another.
}
\details{
This function transfers units from one object to another.
}
