% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biont_help.R
\name{weighted.var.se}
\alias{weighted.var.se}
\title{Compute the Variance of a Weighted Mean}
\usage{
weighted.var.se(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{w}{A numeric vector of weights.}

\item{na.rm}{Logical, if TRUE, NA values are removed before calculation.}
}
\value{
A numeric value representing the standard error of the weighted mean.
}
\description{
Computes the variance of a weighted mean following Cochran 1977 definition.
}
\details{
This function computes the variance of a weighted mean following Cochran 1977 definition.
}
\references{
\insertRef{Gatz1995AE}{BiontR}
}
